MOSAICS Toolbox v1
First draft of toolbox to be completed. Further development needed, but this code is self-contained and should run smoothly on any computer with the required modules.

Next steps include implementing FSLmaths for dilation, brain masking, statistics, measure spreadsheet outputs, and possibly interpretation of Brainsight and raw Signal files to construct the data spreadsheet.